Move simplewebtransport editor into new simplewebtransport.editor asmdef
