+++ Assets/Mirror/Transports/SimpleWeb/SimpleWeb/Client/ClientWebsocketSettings.cs
@@ -1,0 +1,25 @@
+using System;
+
+namespace Mirror.SimpleWeb
+{
+    [Serializable]
+    public struct ClientWebsocketSettings
+    {
+        public WebsocketPortOption ClientPortOption;
+        public WebsocketPathOption ClientPathOption;
+        public ushort CustomClientPort;
+        public string CustomClientPath;
+    }
+    public enum WebsocketPortOption
+    {
+        DefaultSameAsServer,
+        MatchWebpageProtocol,
+        SpecifyPort
+    }
+    public enum WebsocketPathOption
+    {
+        DefaultNone,
+        SpecifyPath
+    }
+
+}
