+++ Assets/Mirror/Transports/SimpleWeb/SimpleWeb/Client/ClientWebsocketSettings.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 20ed869c5ba54a56b750ac9e82be069f
+timeCreated: 1700425326
