+++ Assets/Mirror/Transports/SimpleWeb/SimpleWeb/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a7a134b4d9eef45239a2d7caf7f52c3e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
