+++ Assets/Mirror/Transports/SimpleWeb/SimpleWeb/Editor/SimpleWebTransport.Editor.asmdef
@@ -1,14 +1,12 @@
 {
-    "name": "Mirror.Editor",
+    "name": "SimpleWebTransport.Editor",
     "rootNamespace": "",
     "references": [
+        "GUID:3b5390adca4e2bb4791cb930316d6f3e",
         "GUID:30817c1a0e6d646d99c048fc403f5979",
-        "GUID:1d0b9d21c3ff546a4aa32399dfd33474",
         "GUID:627104647b9c04b4ebb8978a92ecac63"
     ],
-    "includePlatforms": [
-        "Editor"
-    ],
+    "includePlatforms": [],
     "excludePlatforms": [],
     "allowUnsafeCode": false,
     "overrideReferences": false,
