+++ Assets/Mirror/Transports/SimpleWeb/SimpleWeb/Editor/SimpleWebTransport.Editor.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: eb587e0c115d743bfb5e44710c28d003
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
