+++ Assets/Mirror/Transports/SimpleWeb/SimpleWebTransport.cs
@@ -6,26 +6,6 @@ using UnityEngine.Serialization;
 
 namespace Mirror.SimpleWeb
 {
-    [Serializable]
-    public struct ClientWebsocketSettings
-    {
-        public WebsocketPortOption ClientPortOption;
-        public WebsocketPathOption ClientPathOption;
-        public ushort CustomClientPort;
-        public string CustomClientPath;
-    }
-    public enum WebsocketPortOption
-    {
-        DefaultSameAsServer,
-        MatchWebpageProtocol,
-        SpecifyPort
-    }
-    public enum WebsocketPathOption
-    {
-        DefaultNone,
-        SpecifyPath
-    }
-
     [DisallowMultipleComponent]
     public class SimpleWebTransport : Transport, PortTransport
     {
